Fixed db overload

+++ src/main/java/com/sebproject/currency/restfulservice/repo/PeriodRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PeriodRepo extends JpaRepository<CurrencyPeriod, Integer> {
+public interface PeriodRepo extends JpaRepository<CurrencyPeriod, Long> {
 }
